+++ Exodus/src/utopia/exodus/rest/resource/description/LanguageFamiliaritiesNode.scala
@@ -4,6 +4,7 @@ import utopia.citadel.database.access.many.description.DbDescriptions
 import utopia.citadel.database.access.many.language.DbLanguageFamiliarities
 import utopia.exodus.rest.util.AuthorizedContext
 import utopia.metropolis.model.combined.language.DescribedLanguageFamiliarity
+import utopia.metropolis.model.enumeration.ModelStyle.Simple
 import utopia.metropolis.model.stored.description.DescriptionLink
 import utopia.metropolis.model.stored.language.LanguageFamiliarity
 import utopia.nexus.result.Result
@@ -27,6 +28,8 @@ class LanguageFamiliaritiesNode(authorization: (AuthorizedContext, => Result, Co
 		authorization(context, onAuthorized, connection)
 	
 	override protected def items(implicit connection: Connection) = DbLanguageFamiliarities.all
+	
+	override def defaultModelStyle = Simple
 	
 	override protected def descriptionsAccess = DbDescriptions.ofAllLanguageFamiliarities