+++ Metropolis/src/utopia/metropolis/model/combined/description/DescribedDescriptionRole.scala
@@ -13,4 +13,7 @@ object DescribedDescriptionRole extends DescribedFromModelFactory[DescribedDescr
   * @since 20.5.2020, v1
   */
 case class DescribedDescriptionRole(wrapped: DescriptionRole, descriptions: Set[DescriptionLink])
-	extends DescribedWrapper[DescriptionRole]
+	extends DescribedWrapper[DescriptionRole] with SimplyDescribed
+{
+	override protected def simpleBaseModel(roles: Iterable[DescriptionRole]) = wrapped.toModel
+}