+++ Metropolis/src/utopia/metropolis/model/combined/language/DescribedLanguageFamiliarity.scala
@@ -1,7 +1,7 @@
 package utopia.metropolis.model.combined.language
 
-import utopia.metropolis.model.combined.description.{DescribedWrapper, DescribedFromModelFactory}
-import utopia.metropolis.model.stored.description.DescriptionLink
+import utopia.metropolis.model.combined.description.{DescribedFromModelFactory, DescribedWrapper, SimplyDescribed}
+import utopia.metropolis.model.stored.description.{DescriptionLink, DescriptionRole}
 import utopia.metropolis.model.stored.language.LanguageFamiliarity
 
 object DescribedLanguageFamiliarity extends DescribedFromModelFactory[DescribedLanguageFamiliarity, LanguageFamiliarity]
@@ -15,4 +15,7 @@ object DescribedLanguageFamiliarity extends DescribedFromModelFactory[DescribedL
   * @since 25.7.2020, v1
   */
 case class DescribedLanguageFamiliarity(wrapped: LanguageFamiliarity, descriptions: Set[DescriptionLink])
-	extends DescribedWrapper[LanguageFamiliarity]
+	extends DescribedWrapper[LanguageFamiliarity] with SimplyDescribed
+{
+	override protected def simpleBaseModel(roles: Iterable[DescriptionRole]) = wrapped.toModel
+}