+++ Metropolis/src/utopia/metropolis/model/combined/user/DescribedUserLanguage.scala
@@ -1,6 +1,10 @@
 package utopia.metropolis.model.combined.user
 
+import utopia.flow.datastructure.immutable.Constant
+import utopia.flow.generic.ValueConversions._
+import utopia.metropolis.model.combined.description.DescribedSimpleModelConvertible
 import utopia.metropolis.model.combined.language.{DescribedLanguage, DescribedLanguageFamiliarity}
+import utopia.metropolis.model.stored.description.DescriptionRole
 import utopia.metropolis.model.stored.user.UserLanguage
 
 /**
@@ -10,4 +14,9 @@ import utopia.metropolis.model.stored.user.UserLanguage
   */
 case class DescribedUserLanguage(wrapped: UserLanguage, language: DescribedLanguage,
 								 familiarity: DescribedLanguageFamiliarity)
-	extends FullUserLanguageLike[DescribedLanguage, DescribedLanguageFamiliarity]
+	extends FullUserLanguageLike[DescribedLanguage, DescribedLanguageFamiliarity] with DescribedSimpleModelConvertible
+{
+	override def toSimpleModelUsing(descriptionRoles: Iterable[DescriptionRole]) =
+		language.toSimpleModelUsing(descriptionRoles) +
+			Constant("familiarity", familiarity.toSimpleModelUsing(descriptionRoles))
+}