+++ Metropolis/src/utopia/metropolis/model/stored/user/UserSettings.scala
@@ -2,7 +2,8 @@ package utopia.metropolis.model.stored.user
 
 import utopia.flow.datastructure.immutable.{Constant, Model, ModelDeclaration}
 import utopia.flow.generic.ValueConversions._
-import utopia.flow.generic.{FromModelFactoryWithSchema, IntType, ModelConvertible, StringType}
+import utopia.flow.generic.{FromModelFactoryWithSchema, IntType, StringType}
+import utopia.metropolis.model.StyledModelConvertible
 import utopia.metropolis.model.partial.user.UserSettingsData
 import utopia.metropolis.model.stored.Stored
 
@@ -20,8 +21,13 @@ object UserSettings extends FromModelFactoryWithSchema[UserSettings]
   * @author Mikko Hilpinen
   * @since 2.5.2020, v1
   */
-case class UserSettings(id: Int, userId: Int, data: UserSettingsData) extends Stored[UserSettingsData] with ModelConvertible
+case class UserSettings(id: Int, userId: Int, data: UserSettingsData)
+	extends Stored[UserSettingsData] with StyledModelConvertible
 {
-	override def toModel = Model(Vector("id" -> id, "user_id" -> userId, "name" -> data.name,
-		"email" -> data.email))
+	// IMPLEMENTED  --------------------------
+	
+	override def toModel = Model(Vector("id" -> id, "user_id" -> userId)) ++ data.toModel
+	
+	// A simple model omits the user settings version id and uses "id" key for the user id
+	override def toSimpleModel = Constant("id", userId) +: data.toModel
 }